+++ graal/com.oracle.graal.truffle/src/com/oracle/graal/truffle/TruffleSplittingStrategy.java
@@ -1,0 +1,11 @@
+package com.oracle.graal.truffle;
+
+public interface TruffleSplittingStrategy {
+
+    void beforeCall(Object[] arguments);
+
+    void afterCall(Object returnValue);
+
+    void forceSplitting();
+
+}